couldn't make random redraw and contain work. any ideas how?
As seen in the terminal on left mouse button press
keys are changing adding the random key at the end but
the terminals are not drawn in the order that we are setting here.
I know this code looks ugly at this point sorry for that I just wanter
to experiment and see if I could make this work.
Since maps are not ordered I have introduced a new variable of type slice msg.TerminalId to store the
drawing order in terminalstack but it doesn't seem to work.
See implementation changed in terminal/draw.go. Instead of iterating over
the map I iterate over the sorted slice of terminalId keys to access next
terminal to draw but I couldn't make it work as I intended.